+++ lib/gg_hash.dart
@@ -1,6 +1,13 @@
+// @license
+// Copyright (c) 2019 - 2024 Dr. Gabriel Gatzsche. All Rights Reserved.
+//
+// Use of this source code is governed by terms that can be
+// found in the LICENSE file in the root of this package.
+
 /// Support for doing something awesome.
 ///
 /// More dartdocs go here.
 library;
 
 export 'src/fnv1.dart';
+export 'src/hash_json.dart';