+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add a ggHash method and a hashJson method to allow to create 128 bit hashes on arbitrary strings and hash JSON files
+- Add example code
+
 ## [1.0.4] - 2024-04-13
 
 ### Removed
@@ -35,6 +42,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_hash/compare/1.0.4...HEAD
 [1.0.4]: https://github.com/inlavigo/gg_hash/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/inlavigo/gg_hash/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/inlavigo/gg_hash/compare/1.0.1...1.0.2